DE447: Make sure root's SSH key gets generated staged to be shared.
We do the key generation and staging for the admin node in install_chef.sh
instead of in a chef recipe.  This ensures that the root SSH key for the
admin node is added to authorized_keys.erb and staged to be grabbed by
the compute installs before Chef starts managing files.