Fix up some status and node reset handling in test_lib.sh
We don't need to be nice about bringing down the nodes when resetting
them for another series of tests.  Just kill and/or power them off.

Make manually resetting the slaves (by touching the right .reset file
and killing the VM) take less work by having the reset handler in the
main compute node control loop perform the crowbar machines delete.

Clean up final_status handling a bit -- it is still needlessly hairy,
but this is a start.