Add support for OS specific metadata in the barclamps.
This adds support for the crowbar.yml files to declare OS specific
repositories, packages, gems, ppas, and raw packages. The main
crowbar barclamp has been updated to use this functionality, and the
base.list files for all the operating systems in the main crowbar
repository has been reduced to an empty stub.