Add a useful backup function for non github barclamps.
This adds a generic_backup function that makes branches in the
personal/GITHUB_ID namespace that track the local branches outside that
namespace.  This allows you to save your work on non-github barclamps in the
repository from whence the code came.

The barclamp bacing up section of backup_everything has been updated to handle
disambiguating between barclamps that have a personal remote the rest.