Implement a strategy to deal with submodule reference updates.
At pull-requests-gen time, if we have submodule pushes for a branch,
we will create a new commit with all the reference updates added,
push that commit to a throwaway branch in the personal remote,
generate the pull request as normal, and then locally revert the branch back
to the state it was just before adding the new submodule references.

This should avoid any merge conflicts related to submodule reference updates.