This code change attempts to allow the choice of disks for nova-volume while also allowing the system automate the coding.
There is now a type of backing store.  local or raw
local = a local file in the file system capped at the size specified.
The default size is 2TB.  THe system will make sure that the file size
at the time is creation is not more than 90% of the remaining
filesystem space or the specified size.  If no raw disks are found,
this method will be attempted.

By default, the mode is raw and the system will attempt to use all free disks on the
system as specified by the raw_method.  The raw-method can be either all, selected, or
first.  If all is specified, all free disks are used. If selected is specified, the
selected list will be used.  If the selected list is empty, then the first disk available
will be used.  If raw_method is first, then the first disk is used.
If no disks are found, then the local file is built.

On the open source version, the disks will be specified by the user as the node is built.
On the crowbar offering with the raid barclamp, the node will build a raid10 for
redundancy and the local file method will always be defaulted to.  This can be overriden
by the user by manually allocating the node into jbod mode.