Add hierarchy back to builds in releases.
This adds the notion of hierarchy to builds within a release.
If a build has a symlink named "parent" in its directory, then it will
inherit barclamp metadata from the build that the parent symlink points to.

This patch also adds more paranoia when doing builds using a git
managed build cache.