Add "Getting started with Crowbar development" guide
Step by step instructions on how to setup a minimal (Rails only) Crowbar
development environment. Includes steps and scripts to setup the dev VM
with qemu-kvm.

I found it rather tricky and time consuming to setup a minimal
development for poking around the Rails app and related tests. This
document will help others save time by providing one central place for
this, with links to the relevant details in other (existing) docs.

It's still missing a couple of things and will need to be improved on
over time.