[Dev setup] Handle dev setup dependencies in Gemfile
Instead of checking that the required gem dependencies are present and
prompting the user to install it manually if not, specify what we need
in a Gemfile and let bundler take care of it.

This Gemfile should be very small and contain only the depedencies of
`./dev setup-unit-tests`. Everything else is pulled in by the main
Crowbar Gemfile.