[openSUSE] Add docs for setting up an openSUSE dev environment
Now that the Crowbar code is able to run on Ruby 1.9, we can get Crowbar
to run on openSUSE 12.2, which doesn't directly support Ruby 1.8.7 (one
can use `rvm` but we want to avoid adding another layer).

The steps described in the openSUSE guide will only work smoothly when
the following pull requests are merged:

- https://github.com/crowbar/crowbar/pull/1665
- https://github.com/crowbar/crowbar/pull/1666
- https://github.com/crowbar/crowbar/pull/1667
- https://github.com/crowbar/crowbar/pull/1668

Also updated the Ubuntu guide to reflect recent changes and fixes.