Add switches to barclamp_install to skip certain steps
When packaging barclamps into distribution packages (rpm or deb) it is helpful
to split up the barclamp installation process into several steps. So of those
step can be execute during the package build process, like laying out all the
files into the correct directories (e.g. copying chef components into
/opt/dell/chef, copying crowbar app components into
/opt/dell/crowbar_framework, ...).
Other step need to run inside the installed system (upload of cookbooks,
database migrations, ...). This patch introduces the following new commandline
switches for that:
--no-migrations: To not run execute the database migrations
--no-framework-install: To skip the basic step of laying out the base files in
the correct subdirectory (use this on the installed system, when calling
barclamp_install for a barclamp that was installed from a packages.