dev: don't relocate Gemfile
61a43b1c copied the Gemfile to the root of the test installation, and
1c60c571 changed the copy to a move.  However, this relocation breaks
Rails conventions, doesn't match what happens outside the test
environment, and creates extra complications e.g. with the rsync in
travis-ci/update-git.sh, and introduces uncertainty over whether
Gemfile.lock and Gemfile.d should also be relocated or not.

It's perfectly easy to ensure that you're in the correct directory
before running bundle exec, or simply passing --gemfile=... to bundle,
so let's take this conventional approach instead, and keep the Gemfile
within the Rails app to which it's coupled.