Guard config to automatically push changes from git checkout
When Guard is running with this config, it will watch changes in each
barclamp and push all relevant changes under `crowbar_framework/` to the
dev-test directory.

This allows changes in the git checkout to immediately show up in your
local Crowbar instance that is running in development mode - essential
when working on the Crowbar frontend.

This is just a stop-gap solution for me and Ken as we hack on the
frontend. Adam is meanwhile working on a more long-term and flexible
solution that also triggers the unit tests.