Revert "Guard config to automatically push changes from git checkout"
This reverts commit 1b5ea996146e3b9f735fd7e6308b856427c6316e.

As Victor pointed out, this commit apparently breaks the dev tool which
assumes that everything in barclamps/ is a barclamp (regardless if it is
a directory or file). We'll have to patch the dev tool in several places
to fix this, so for now it's easier to move these files elsewhere.

Removing them for now until I get it to run from dev-setup/.