install-chef-suse: Add infrastructure for improved output/logging
We don't have to put everything on stdout; it would be more friendly to
only output a summary on stdout, and keep the full output in the log
file.

Ther CROWBAR_VERBOSE environment variable can be set to disable this.

Part of the code (especially the spinner) is based on similar code from
DevStack.