Add support for git tracked metadata.
This is intended to let other people test drive using git tracked
metadata.  Support is not complete -- the dev tool still needs to be
updated to pull in and track the metadata tracking repo like it tracks
the barclamps, and support needs to be added to ./dev setup to start
pulling in the metadata tracking repo.