Added ```#!/usr/bin/env bash``` to dev script
In order to execute the dev script on other operating systems the env
decision for the bash binary should be fine.

In a specific case i tried
to execute the dev script on a Mac dev machine. As the system bash is
quite outdated i installed a newer bash version that is not localted at
```/usr/bin``` and i was not able to execute the bash script. Even
```/usr/local/bin/bash dev``` won't work as the "parameters" are getting
parsed wrong.