Reduce flickering caused by the spinner a bit
Normally  dialog  checks  to  see if it is running in an xterm, and in that case
tries to suppress the initialization strings that would make it  switch  to  the
alternate screen.  Switching between the normal and alternate screens is visual-
ly distracting in a script which runs dialog several times.