Stop excessive chef-server log
The logs grow to ~200MB in a couple of hours when we deploy, because it
logs the full params (eg, full content of node objects) at the info
level. The log is not really useful as it is, so instead, only log at
the warn level.

https://bugzilla.novell.com/show_bug.cgi?id=884550

Workaround mentioned in original upstream bug:
 https://tickets.opscode.com/browse/CHEF-1122