Fixed tftpboot symlink check
As it is possible to use autofs to mount the medias we need to set
symlinks to the autofs root folder. Now the condition checks if the
symlink is set within the tftpboot root, if it is everything is fine, if
it points outside of the root the installation stops with an error
message.