Set frame_max setting of rabbitmq to 0 during install (bnc#910815)
The versions of bunny prior to 0.9 don't support frames; but chef
requires version < 0.8. So work around this by disabling frames.

As we know that later versions of chef won't require this, also add a
"reminder crash" to revert this when we have a more recent version of
chef.

Note that we do this as an environment variable, because this is really
just our way to patch the rabbitmq configuration, not to set it forever
like this.

http://bugzilla.suse.com/show_bug.cgi?id=910815