Remove ./dev tool, build scripts, packaging scripts, test framework
All these bits are deeply connected to each other, and since we don't
use the ./dev tool anymore, let's just drop everything.

The build script was used to create an ISO with everything, but we want
to use packages instead.

The packaging should be done downstream.

The test framework is now moved to mkcloud and other similar scripts.