Move sanity checks to separate crowbar lib
This change removes all network checks except the network-json-validator
test. The crowbar service starts before we create the proposals and the
tests are executed at every crowbar start. That means that we don't lose
test coverage. The network-json-validator stays for the moment because
it makes no sense to install crowbar with an broken/wrong network json.
We can drop this test when we start install-chef-suse from crowbar
itself.