Refactor running rake (bsc#993035)
If barclamp_install fails to upload cookbooks, it does not check the
exit status of the bin/rake call, and merrily continues on, finally
exiting with status 0. Refactor the code that calls bin/rake to
no longer make use of system() or su, and instead use spawn(). This
also means that output from bin/rake chef:upload:all is also logged,
it wasn't previously.

Also clean-up the begin/rescue block when running schema migrations,
the function exits if there is an error, and doesn't throw an
exception.