Set no_proxy in install-chef-suse.sh (bsc#998968)
Chef now works around a missing Ruby feature by checking no_proxy
itself and matching against it when returning if a proxy connection
is required, so drop the http_proxy workaround and set no_proxy
directly. Since the script talks to both the hostname and the IP, add
them both.