upgrade: No need to use crowbar-init during the upgrade
crowbar-init is used to bootstrap crowbar. This has been already
done as part of the installation and we do not need to do it again
during upgrade.
In 7-8 upgrade process, we're not changing crowbar database,
so crowbar process could simply start with existing postgresql backend
after the admin server upgrade and reboot.