Add IPv6 support to a network-json-validator and install-chef-suse.sh
This patch add IPv6 support to network-json-validator and the
install-chef-suse.sh scripts as required to run an IPv6 CP.

The network-json-validator tool now supports sending the netmask in
in cidr from, eg 32 or 64 etc. And also supports all the IPs elements
as ipv6 addresses.

If the netmask is a CIDR then it also makes sure the value is a valid
integer value for the IP address type:

  IPv4 - 1-32
  IPv6 - 1-128

It determine's the IP type based of the subnet address.

CIDR's are then converted to the full netmask representation
so it can get all the normal netmask checks.

The install-chef-suse.sh script likewise supports the admin ip being
ipv6 and will also wrap the IP in '[]' when required.
As well as making sure that when using IPv6 the chef-sever listens
for IPv6 connections by listening on [::].

If the admin ip is an ipv6 address this patch will place an
/etc/sysconfig/chef-server file (as the systemd service expects)
to provide other options, in our case it adds:

  OPTIONS="-h ::"