install-chef-suse: Give a hint to the user when bootstrap wasn't run
While it isn't entirely clear to my why we separated bootstrap
from install, the error handling of missing bootstrap is suboptimal.

Detect the case that bootstrap failed prior trying to continue
with the installation, as it will never succeed and the error messages
are very hard to understand for the occassional user.