ipv6: Ensure ip version uniformatiy across networks
Ensure all networks are using the same address family and fail
validation if the user is attempting to mix and match IPv4 and
IPv6 networks.

In the early versions of IPv6 support to crowbar, we want to
minimalise the amount of extra testing would be required by
CI/CD. So as an initial step we will only support all IPv4 or
all IPv6 networks. This change inforces that in the
network-json-validator