crowbar: Relax conditional for aborting installation (SCRD-8330)
The `install-chef-suse.sh` script is not only called during initial
deployment, but also during a Crowbar restore operation. During a
restore, the `crowbar-init.service` might not be running and the
current conditional testing for whether the installation can continue
is too strict in this case. This change relaxes the conditional by
unconditionally running the `reset_crowbar` step, while intercepting a
potential failure.

Signed-off-by: Nicolas Bock <nicolas.bock@suse.de>