ipv6: Update start-up.sh to be IPv6 compatible
The slegehammer/start-up.sh script that is used in the sleshammer
discovery image only supports IPv4 addresses.

This change adds the required IPv6 support. The most notible addition
is how we determine the admin IP address so we can mount the nfs mounts,
etc.
In IPv4 we use wicked's wickedd-dhcp4 binary to get network information
and use the returned SERVERID as the admin ip.
However when using wickedd-dhcp6 the SERVERID ends up just being
something that looks similar to the DUID for the server, NOT the IP.
The only other IP's that could gleaned from this return output is the DNS
address or to grab the pxe boot url and try and grab the host out of that.
But I worry there would be deployments where all these services would be handled
by other servers.

So in order to obtain the admin ip, this patch adds an optional extra
cmdline option 'adminip' which can be added to the kernel commandline.
If there this script has been modified to use that, otherwise it falls
back to the SERVERID.

I have modified the provisioner barclamp, in crowbar-core, to append the
'adminip=' option to the grub2.cfg and pxelinux.cfg/default files for
the discovery image. And will push the PR as soon as this patch has
landed.

I've also branched the cloud9 sleshammer package in IBS and updated it with
this version of the start-up.sh script, and again will submitpac to staging
as soon as this lands.

I have tested the branched sleshammer image with the updated start-up.sh
file on my local mkcloud instance on my laptop.