install-chef-suse: Do not depend on machine-install user
With the goal of removing the machine-install in mind, this commit moves
install-chef-suse to:

  - create a /etc/crowbarrc file
  - use crowbarctl instead of crowbar CLI

Note that the "crowbarctl proposal show" call needs --raw, as otherwise
crowbarctl tries to be clever, but fetches the list of clusters -- which
is not possible during the install process (that API is not up yet).

We still create the machine-install user as other bits in Crowbar still
depend on it, but we'll eventually remove it.