IPv6: Export ip_version and handle the DHCP domain
The sleshammer image runs the start-up.sh script. This bootstraps
the discovery image on the node and then passes to crowbar-core's
control.sh, via nfs, to finish bootrapping the node.
The control.sh script needs to know the IP verson, start-up.sh knows
what it is but this isn't passed on to control via the environment.

This patch exports the IP_VERSION and ADMIN_IP_WRAPPED variables into
the env so control.sh can pick them up and do the right the righ thing.

The wicked dhcp6 client doesn't generate a DNSDOMAIN variable, so
start-up.sh fails to generate a correct FQDN. In dhcpd6 you can
use the DNS Search domains option which is picked up as DNSSEARCH in wicked.

To support IPv6, if the DNSDOMAIN is missing the script will look for
DNSSEARCH and use the first domain in the list. Note in crowbar we'll only
be passing in one domain. If that fails instead of creating an incorrect
FQDN the full hostname will only contain the short hostname,
ie missing the domain.